포인트수정

+++ front-end/src/components/housework/HouseWork2.jsx
@@ -4,7 +4,7 @@ import axios from "axios";
 import { useSelector } from "react-redux";
 import WorkModal from "./WorkModal";
 import CompleteModal from "./CompleteModal";
-import { BsThreeDotsVertical, BsPlusCircle } from "react-icons/bs";
+import { BsThreeDots, BsPlusCircle } from "react-icons/bs";
 import { FcRating } from "react-icons/fc";
 import Modal from "react-modal";
 
@@ -221,20 +221,23 @@ const Housework2 = () => {
 
   // 선택한 작업 삭제하는 함수
   const deleteSelectedTasks = async (workIdx, taskType) => {
-    try {
-      await axios.delete(
-        `http://localhost:8089/wefam/delete-work/${workIdx}`
-      );
+    if (window.confirm("할 일을 삭제하시겠습니까?")) {
+      try {
+        await axios.delete(
+          `http://localhost:8089/wefam/delete-work/${workIdx}`
+        );
 
-      setTasks((prevTasks) => ({
-        ...prevTasks,
-        [taskType]: prevTasks[taskType].filter((task) => task.workIdx !== workIdx),
-      }));
+        setTasks((prevTasks) => ({
+          ...prevTasks,
+          [taskType]: prevTasks[taskType].filter((task) => task.workIdx !== workIdx),
+        }));
 
-      fetchTasks();
-    } catch (error) {
-      console.error("작업 삭제 중 오류 발생:", error);
+        fetchTasks();
+      } catch (error) {
+        console.error("작업 삭제 중 오류 발생:", error);
+      }
     }
+
   };
 
 
@@ -364,6 +367,9 @@ const Housework2 = () => {
           <br />
           <span>{task.workContent}</span>
           <div className={styles.userContainer}>{renderCompletedTaskUsers(task)}</div>
+          <p className={styles.completedDate}>
+            완료일: {task.completedAt ? task.completedAt.toLocaleDateString() : "완료일 정보 없음"}
+          </p>
         </div>
         <div className={styles.dropdownContainer}>
           <div className={styles.taskRight}>
@@ -417,7 +423,7 @@ const Housework2 = () => {
           <span className={styles.taskPoints}>{task.points} 포인트</span>
 
           {!isCompleted && (
-            <BsThreeDotsVertical
+            <BsThreeDots
               className={styles.taskIcon}
               onClick={(e) => {
                 e.stopPropagation();
